ushort changed to uint for operations related to sample rate. This is done as we are now supporting 96000 Hz sample rate.

+++ csharp/daisy/export/Daisy3_Export_ImageDescription_BodyContent.cs
@@ -1213,7 +1213,7 @@ namespace urakawa.daisy.export
                     {
                         dataProvider.ExportDataStreamToFile(destPath, false);
                         if (encodeToMp3 ||
-                            (ushort)pcmFormat.SampleRate != (ushort)managedAudio.AudioMediaData.PCMFormat.Data.SampleRate)
+                          (uint)pcmFormat.SampleRate != (uint)managedAudio.AudioMediaData.PCMFormat.Data.SampleRate)
                         {
                             string convertedFile = EncodeWavFileToMp3(skipACM, destPath, encodeToMp3, pcmFormat,
                                                                       bitRate_Mp3);
